Move preferences into a separate file
